Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

So I updated this so that if the kernel version-specific directory exists, it will load modules from there, but if not, it will attempt to load from the base module directory.

This should get modprobe working on devices that don't put kernel modules in the kernel version-specific directory while also not breaking modprobe on devices that do.

Patch-set: 2