Enable nfs mount in both dynamic and static busybox.
With this change, busybox is able to do NFSv3 mounts, if the kernel
has been compiled with NFS client support.

While busybox already includes the neccessary code in
util-linux/mount.c, the required SUN RPC library was missing. It is
included here in the librpc/ subdirectory. The librpc/ code was taken
from http://nbd.name/gitweb.cgi?p=uclibc-rpc.git;a=summary and modified
a bit, mostly in librpc/rpc/types.h, to properly compile for Android/CM9.
I also compared the code to the current official uclibc GIT - it was
almost identical, except for a funny thread support thing in the
official uclibc sources which wouldn't have worked, and a small
fix from there which I incorporated in clnt_{tcp,udp,unix}.c,
where a break was missing from a case.

Tested on a maguro build of the current CM9, together with a kernel
out of codeworkx git / experimental branch, with NFS enabled.

Change-Id: I587ed7c60aa7dd10828213ae2b8675579d126e8e