Update patch set 3
Patch Set 3:

BTW, contrary to what the commit message, accidentally not amended in that place, says: my recent tests were with the kernel for GNex as released by kalimochoAz, both 3.0.22 and 3.0.23 versions, recompiled to have NFS client support built in - i.e. it works with the kernel that CM9 "officially" has on the galaxy nexus.

I think it would be useful, should this change to busybox be merged, to convince kalimochoAz to enable NFSv3 client support in the "official" kernel, too. So, please give me a wink whether this change here would be merged, and I'll talk to him about that.

Patch-set: 3