Update patch set 6
Patch Set 6: No score

They can't both be kept, the default behavior of CMs busybox umount is to detach loop devices on umount, the -D flag tells it not to. This patch sets its default behavior NOT to detach on umount, and -d to make it detach.

You could keep the -D option, but it would be the same as w/o an option at all. If that's what you mean, then it could be that way relatively easy and I could upload another patchset.

Patch-set: 6