Update patch set 7
Patch Set 7:

they added back in april, to be in line with util-linux mount/umount, though their implementation is different and wrong (IMO of course). util-linux umount will automatically detach the loop device by default, if it detects loop= in the mtab for that particular device, meaning the loop device was created with mount, not losetup. BB umount is blindly detaching all loop devices not just the ones its creating. I would patch that implementation here, but it would be pointless as android doesnt have proper mtab support, just a symlink to /proc/mounts

Patch-set: 7