Update patch set 1
Patch Set 1:

Currently I'm just trying to get x86 to build so that I can run CM in the emulator at reasonable speeds.  Adding the x86 stubs is necessary to get the x86 build to complete, but may not be sufficient to get busybox to actually run on x86.  If more work is required to get it to run, I'd prefer those changes to be made in a separate commit.

Patch-set: 1