Update patch set 1
Patch Set 1:

> jellybean branch is already working for x86

This change is about refreshing the MIPS and ARM stubs -- it is unrelated to x86.

My questions for you:
- Do you want the busybox stubs on the other branches to match cm-10.1, or do you want the stubs to match the bionic stubs for that branch?
- Are you holding off on merging this change until refreshes for all branches have been uploaded?

Note that Change-Id: Ia14da35acbc36aafd44a059146a37271bd358ae2 <http://review.cyanogenmod.org/32334> is a prerequisite for this change.  If that change is not merged before this one, MIPS support will break due to the wrong syscall number for getsid().

Patch-set: 1