Update patch set 1
Patch Set 1:

> well, a merge is also possible, but the incoming 1.21.0 commit include half of the changes...

All the more reason to merge jellybean into cm-10.1.  :)  (by merging first, it'll make it easier to see what changed with the new upstream busybox, should simplify merge conflict resolution, etc.)

In general, it's not a good idea for there to be commits on an older branch (jellybean) that aren't also reachable from the newer branch (cm-10.1) -- that's a recipe for regressions and duplicated work (such as me going through the trouble of generating x86 stubs when jellybean already had them).  Cherry-picking the changes across branches (which is what this commit basically does) isn't a good idea because it makes it hard to locate the true origin of a change.  That, in turn, makes it hard to identify which branches are affected by a particular change, which makes it hard to identify all releases affected by a security vulnerability, etc.

> The only remaining diff is the Android.mk cross compiler for x86

The stubs themselves are different.  The stubs in this commit are copied directly from the bionic directory; the stubs on jellybean were modified in commit 3eb78254661e1ffd893dab3b532ca9e31f9161ec (change-id Ic1bd17611abb74c0cfa963a5967b8c21166bcc08).

I think it'd be both cleaner and easier to simply merge jellybean into cm-10.1.

Patch-set: 1