Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

What's the change in include/platform.h about?  It's not related to CROSS_COMPILER_PREFIX, and it's not mentioned in the commit message.

If this commit is a necessary step toward adding x86 support, perhaps it would be best to squash it with <http://review.cyanogenmod.org/32336>.

Patch-set: 1