Update patch set 2
Patch Set 2:

Currently, there is no error message as all. It automatically falls back to mounting read-only, which is successful.

The main impact is that quite a few tools need to mount e.g. /system read-write. This no longer works with the command "mount -o remount,rw /system" as it used to. The "blockdev" command is not available everywhere, plus it doesn't accept the moint point (/system) as parameter but only the device (/dev/...). So I don't think that's a good workaround.

Patch-set: 2