Update patch set 1
Patch Set 1:

Yes, scontext is null. printf probably copes with null when %s is used, but in this case the pointer is dereferenced before it is given to printf.

I just added the same if which is used in other parts of the function (and which also use %lc to print the context).

Patch-set: 1