Define a writeable CONFDIR for iproute2.
Currently, the CONFDIR defaults to /etc/iproute2, which is on a read-only
filesystem. We plan to write some files at runtime to this directory, so we need
to point it at a read-write filesystem.

Note that nothing in Android actually uses (populates) the current CONFDIR
(/etc/iproute2). It doesn't even exist on the default system image. In other
words, this change doesn't break anything.

Change-Id: I69deb1a64d5d6647470823405bf0cc55b24b22de
JIRA: CYAN-6759