dmesg: rework color feature, compat. with more
when used with less or more, the colors were not in
the same buffer, resulting blocs of colors.

Also, reset color numbers to the 16 ANSI ones.

Reduce the diff with busybox to be able to send the patch.
This is the first full busybox aosp feature with the new build system

Signed-off-by: Tanguy Pruvot <tanguy.pruvot@gmail.com>

Change-Id: Ia0239a0bea02da01cae15736150dfee4fe66dcdb