Update patch set 2
Patch Set 2:

Do you mind explaining that to me? My knowledge it not that great in busybox. Because, why is "undef stpcpy" needed for arm/mips and not for x86. And i can't get a clean build in cm-11.0 (4.4.4) for my 2 x86 devices (tab 3 10.1 and razr i). While when switching it back i can build for both just fine like in cm-10.2 (4.2.2).

Or stpcpy needs to be out commented for x86 within the busybox code. Because u can't build clean without.

Patch-set: 2