Update patch set 1
Patch Set 1:

This should fix my stupid logic derp, but i'm somewhat on the fence about what was mentioned in on the mailing list about auto and the user specifying -t.

If anything, for the -t, we should maybe just have a check if its specified and bail out if blkid disagrees.

If blkid KNOWS what the partition type is, and we attempt to specify something different, I don't think we should bypass blkid.

Patch-set: 1