Update patch set 2
Patch Set 2:

In that gist, lines 7/8 actually appears an upstream busybox issue (bug?)

Even if you specify a filesystem type on the command line, if it has to go to fstab to get your device name, it's going to also use fstab's filesystem type.

Not sure if that's expected behavior or not. =/

Patch-set: 2