Update patch set 2
Patch Set 2:

Know we're going the way of Toybox, but wanted to get this updated as it's still used in a number of projects. Feedback encouraged on the methods/changes.

Due to amount of changes, merged using 1.26.2 as base and applied diffs from this repo to 1.22.1 on top.
Made new configs with settings from previous, enabling all new features in full and none in minimal. Links updated accordingly too.

Couple of notes for some sticking spots:
archival/gzip.c: ffs is located in strings.h for Bionic vs string.h, so added an include
Config:
- CONFIG_ASH_INTERNAL_GLOB is y as Bionic doesn't provide this
- CONFIG_FEATURE_SYNC_FANCY is n as Bionic doesn't provide the syscall
- NSENTER & UNSHARE is n as it requires _GNU_SOURCE set, which this project does not have set as far as I can tell
- shell/ash.c - changed evaltreenr to void from int as Clang 3.8 segfaults on this (works on 4.0+). It's a no return function, so can't see a problem with that

Patch-set: 2